BOOKING-202: Moved create booking into own component
